New version 3.1 of the Delete Fabric Records Add-in
Added a new tool for deleting a selection of connection lines. This code
implements a forward star to trace the downstream connection lines
belonging to the same parcel (see the function TraceFabricLines, and its
use of the IGSForwardStar interface.) The selected connection lines are
deleted. If there are non-connection lines selected, they are not
deleted. If there are any connection lines that are downstream from any
of the selected connection lines, then these are also deleted. This
ensures that there is no instance of connection lines that are
disconnected from the parcel.