Added Sample Parcel Trace tool
Added Sample Parcel Trace tool. Demonstrates how to leverage the
TraceConstructor class and make a tool that leverages the tracing
functionality, and the edit sketch, for use with the fabric. In this
code sample the sketch geometry is used to search the fabric boundary
lines and returns a report with information about the geometry of
different types of lines found.