Added code for parcel locking and regenerate
Added code for parcel locking and regenerate. In some cases
ICadastralFabric :: AdjustFabric method does not regenerate the parcels,
so added a regenerate on the parcels affected by directly referenced
line points being moved. For the cases of merge points the point id
references are updated. These need a parcel lock, so code added for the
parcels affected by those types of edits.