Installed new option for Line Point management
New option for line point management. The added option allows the user
to by-pass the automatic updates for un-referenced line-points. This
means that when this option is chosen, the line-points need to be
directly referenced and the new line point location computed for the new
line's position. This gives more control over line-point placement, but
also requires additional work to calculate the positions for the
additional reference features.